Implemented updated group service to handlers

+++ kmaooad-functions/src/main/java/edu/kmaooad/handler/impl/group/assign/AssignActivitiesToGroupHandler.java
@@ -2,7 +2,6 @@ package edu.kmaooad.handler.impl.group.assign;
 
 import edu.kmaooad.constants.bot.ConversationState;
 import edu.kmaooad.constants.bot.GroupConstants;
-import edu.kmaooad.handler.ButtonRequestHandler;
 import edu.kmaooad.handler.UserRequestHandler;
 import edu.kmaooad.helper.KeyboardHelper;
 import edu.kmaooad.model.Activity;
@@ -11,10 +10,10 @@ import edu.kmaooad.model.HandlerResponse;
 import edu.kmaooad.model.UserRequest;
 import edu.kmaooad.service.ActivityService;
 import edu.kmaooad.service.GroupService;
+import edu.kmaooad.service.ServiceException;
 import edu.kmaooad.service.TelegramService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
-import org.telegram.telegrambots.meta.api.objects.replykeyboard.ReplyKeyboardMarkup;
 
 import static edu.kmaooad.constants.bot.GroupConstants.ASSIGNED;
 import static edu.kmaooad.constants.bot.GroupConstants.GROUP_MAP_KEY;
@@ -43,20 +42,27 @@ public class AssignActivitiesToGroupHandler implements UserRequestHandler {
         return UserRequestHandler.isTextMessage(userRequest.getUpdate()) &&
                 userRequest.getUserSession().getConversationState().equals(ConversationState.WAITING_FOR_GROUP_ACTIVITY_ASSIGN_CHOICE);
     }
-    //todo
-    // take off comment
+
     @Override
     public HandlerResponse handle(UserRequest userRequest) {
         Group group = (Group) userRequest.getUserSession().getData().get(GroupConstants.GROUP_MAP_KEY);
         String text = userRequest.getUpdate().getMessage().getText();
-       /* if(text.contains(ASSIGNED)) {
-            String activityName = text.replace(ASSIGNED, "").trim();
-            Activity activity = activityService.getActivityByName(activityName);
-            groupService.deleteActivityGroup(activity.getId(), group.getId());
-        } else {
-            Activity activity = activityService.getActivityByName(text);
-            groupService.addActivityGroup(activity.getId(), group.getId());
-        }*/
+
+        try {
+            if (text.contains(ASSIGNED)) {
+                String activityName = text.replace(ASSIGNED, "").trim();
+                Activity activity = activityService.getActivityByName(activityName);
+                groupService.deleteActivityGroup(activity.getId(), group.getId());
+                log.info("Deleted activity " + activityName + " from group " + group.getName());
+            } else {
+                Activity activity = activityService.getActivityByName(text);
+                groupService.addActivityGroup(activity, group.getId());
+                log.info("Added activity " + activity.getName() + " to group " + group.getName());
+            }
+        } catch (ServiceException e) {
+            e.printStackTrace();
+            log.error("Failed to delete or add activity " + text + " to group" + group.getName());
+        }
 
         userRequest.getUserSession().getData().put(GROUP_MAP_KEY, groupService.getGroupById(group.getId()));
         userRequest.getUserSession().setConversationState(ConversationState.WAITING_FOR_GROUP_TO_ASSIGN_CHOICE);
